requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
python-docx
chardet
requests
streamlit
spacy
